More code smells
