+++ projects/ccf-eui/src/app/app.component.ts
@@ -10,8 +10,7 @@ import { DataQueryState, DataState } from './core/store/data/data.state';
 import { FiltersPopoverComponent } from './modules/filters/filters-popover/filters-popover.component';
 import { ImageViewerPopoverComponent } from './modules/image-viewer/image-viewer-popover/image-viewer-popover.component';
 import { DrawerComponent } from './shared/components/drawer/drawer/drawer.component';
-import { OrganInfo } from '../../../ccf-shared/src/lib/components/organ-selector/organ-selector.component';
-import { ALL_ORGANS } from '../../../ccf-rui/src/app/core/store/model/model.state';
+import { ALL_ORGANS } from '../../../ccf-shared/src/lib/components/organ-selector/organ-selector.component';
 
 /**
  * This is the main angular component that all the other components branch off from.
