+++ projects/ccf-rui/src/app/core/store/model/model.state.ts
@@ -74,21 +74,21 @@ export interface ModelStateModel {
   extractionSets: ExtractionSet[];
 }
 
-export const ALL_ORGANS: OrganInfo[] = [
-  { src: 'app:colon', name: 'Colon', hasSides: false, hasSex: true },
-  { src: 'app:heart', name: 'Heart', hasSides: false, hasSex: true },
-  { src: 'app:kidney', name: 'Kidney', hasSides: true, hasSex: true },
-  { src: 'app:spleen', name: 'Spleen', hasSides: false, hasSex: true },
-  { src: 'app:bladder', name: 'Bladder', disabled: true, hasSides: false, hasSex: true },
-  { src: 'app:brain', name: 'Brain', disabled: true, hasSides: false, hasSex: true },
-  { src: 'app:liver', name: 'Liver', disabled: true, hasSides: false, hasSex: true },
-  { src: 'app:lung', name: 'Lung', disabled: true, hasSides: true, hasSex: true },
-  { src: 'app:lymph_nodes', name: 'Lymph Nodes', disabled: true, hasSides: false, hasSex: true },
-  { src: 'app:ovaries', name: 'Ovaries', disabled: true, hasSides: true, hasSex: false },
-  { src: 'app:small_intestine', name: 'Small Intestine', disabled: true, hasSides: false, hasSex: true },
-  { src: 'app:stomach', name: 'Stomach', disabled: true, hasSides: false, hasSex: true },
-  { src: 'app:thymus', name: 'Thymus', disabled: true, hasSides: false, hasSex: true }
-];
+// export const ALL_ORGANS: OrganInfo[] = [
+//   { src: 'app:colon', name: 'Colon', hasSides: false, hasSex: true },
+//   { src: 'app:heart', name: 'Heart', hasSides: false, hasSex: true },
+//   { src: 'app:kidney', name: 'Kidney', hasSides: true, hasSex: true },
+//   { src: 'app:spleen', name: 'Spleen', hasSides: false, hasSex: true },
+//   { src: 'app:bladder', name: 'Bladder', disabled: true, hasSides: false, hasSex: true },
+//   { src: 'app:brain', name: 'Brain', disabled: true, hasSides: false, hasSex: true },
+//   { src: 'app:liver', name: 'Liver', disabled: true, hasSides: false, hasSex: true },
+//   { src: 'app:lung', name: 'Lung', disabled: true, hasSides: true, hasSex: true },
+//   { src: 'app:lymph_nodes', name: 'Lymph Nodes', disabled: true, hasSides: false, hasSex: true },
+//   { src: 'app:ovaries', name: 'Ovaries', disabled: true, hasSides: true, hasSex: false },
+//   { src: 'app:small_intestine', name: 'Small Intestine', disabled: true, hasSides: false, hasSex: true },
+//   { src: 'app:stomach', name: 'Stomach', disabled: true, hasSides: false, hasSex: true },
+//   { src: 'app:thymus', name: 'Thymus', disabled: true, hasSides: false, hasSex: true }
+// ];
 
 /**
  * Data for the main 3d model display
