+++ projects/ccf-shared/src/lib/components/organ-selector/organ-selector.component.ts
@@ -1,6 +1,22 @@
 import { AfterViewInit, OnDestroy, Component, EventEmitter, HostBinding, Input, Output, ElementRef, ViewChild } from '@angular/core';
 import { ResizeSensor } from 'css-element-queries';
 
+export const ALL_ORGANS: OrganInfo[] = [
+  { src: 'app:colon', name: 'Colon', hasSides: false, hasSex: true },
+  { src: 'app:heart', name: 'Heart', hasSides: false, hasSex: true },
+  { src: 'app:kidney', name: 'Kidney', hasSides: true, hasSex: true },
+  { src: 'app:spleen', name: 'Spleen', hasSides: false, hasSex: true },
+  { src: 'app:bladder', name: 'Bladder', disabled: true, hasSides: false, hasSex: true },
+  { src: 'app:brain', name: 'Brain', disabled: true, hasSides: false, hasSex: true },
+  { src: 'app:liver', name: 'Liver', disabled: true, hasSides: false, hasSex: true },
+  { src: 'app:lung', name: 'Lung', disabled: true, hasSides: true, hasSex: true },
+  { src: 'app:lymph_nodes', name: 'Lymph Nodes', disabled: true, hasSides: false, hasSex: true },
+  { src: 'app:ovaries', name: 'Ovaries', disabled: true, hasSides: true, hasSex: false },
+  { src: 'app:small_intestine', name: 'Small Intestine', disabled: true, hasSides: false, hasSex: true },
+  { src: 'app:stomach', name: 'Stomach', disabled: true, hasSides: false, hasSex: true },
+  { src: 'app:thymus', name: 'Thymus', disabled: true, hasSides: false, hasSex: true }
+];
+
 /**
  * Contains the organ name and url of the icon svg
  */
@@ -54,7 +70,7 @@ export class OrganSelectorComponent implements AfterViewInit, OnDestroy {
   /**
    * List of organs in the carousel
    */
-  @Input() organList: OrganInfo[] = [];
+  @Input() organList: OrganInfo[] = ALL_ORGANS;
 
   /**
    * Currently selected organ
